requirements.txt
Flask
mysql-connector-python
flask-cors
flask-jwt-extended
bcrypt
pyodbc
gunicorn
dotenv
google.generativeai